python_iot/requirements.txt
newrelic
python-dotenv
sinricpro==5.2.1
requests